refactor

+++ src/host_mesh.cpp
@@ -22,7 +22,8 @@ host_mesh_t::host_mesh_t( const std::vector<glm::vec3> &vertices,
     aabb_cached_( false ),
     aabb_( std::make_pair( glm::vec3( std::numeric_limits<float>::max() ), glm::vec3( std::numeric_limits<float>::min() ) ) ) {}
 
-host_mesh_t::host_mesh_t( const tinyobj::shape_t &shape ) {
+host_mesh_t::host_mesh_t( const tinyobj::shape_t &shape )
+{
     std::vector<glm::vec3> vertices;
     std::vector<glm::vec3> normals;
     std::vector<glm::vec2> texcoords;
@@ -32,12 +33,14 @@ host_mesh_t::host_mesh_t( const tinyobj::shape_t &shape ) {
     glm::vec3 diffuse_color;
 
     {
-        int totalsize = shape.mesh.indices.size() / 3;
+        int totalsize = int(shape.mesh.indices.size() / 3);
         int f = 0;
-        while( f < totalsize ) {
+        while( f < totalsize )
+        {
             int process = std::min( 10000, totalsize - f );
             int point = 0;
-            for( int i = f; i < process + f; i++ ) {
+            for( int i = f; i < process + f; i++ )
+            {
                 int idx0 = shape.mesh.indices[3 * i];
                 int idx1 = shape.mesh.indices[3 * i + 1];
                 int idx2 = shape.mesh.indices[3 * i + 2];
@@ -55,7 +58,8 @@ host_mesh_t::host_mesh_t( const tinyobj::shape_t &shape ) {
                 vertices.push_back( p1 );
                 vertices.push_back( p2 );
 
-                if( shape.mesh.normals.size() > 0 ) {
+                if( shape.mesh.normals.size() > 0 )
+                {
                     normals.push_back( glm::vec3( shape.mesh.normals[3 * idx0],
                                                   shape.mesh.normals[3 * idx0 + 1],
                                                   shape.mesh.normals[3 * idx0 + 2] ) );
@@ -65,21 +69,26 @@ host_mesh_t::host_mesh_t( const tinyobj::shape_t &shape ) {
                     normals.push_back( glm::vec3( shape.mesh.normals[3 * idx2],
                                                   shape.mesh.normals[3 * idx2 + 1],
                                                   shape.mesh.normals[3 * idx2 + 2] ) );
-                } else {
+                }
+                else
+                {
                     glm::vec3 norm = glm::normalize( glm::cross( glm::normalize( p1 - p0 ), glm::normalize( p2 - p0 ) ) );
                     normals.push_back( norm );
                     normals.push_back( norm );
                     normals.push_back( norm );
                 }
 
-                if( shape.mesh.texcoords.size() > 0 ) {
+                if( shape.mesh.texcoords.size() > 0 )
+                {
                     texcoords.push_back( glm::vec2( shape.mesh.positions[2 * idx0],
                                                     shape.mesh.positions[2 * idx0 + 1] ) );
                     texcoords.push_back( glm::vec2( shape.mesh.positions[2 * idx1],
                                                     shape.mesh.positions[2 * idx1 + 1] ) );
                     texcoords.push_back( glm::vec2( shape.mesh.positions[2 * idx2],
                                                     shape.mesh.positions[2 * idx2 + 1] ) );
-                } else {
+                }
+                else
+                {
                     glm::vec2 tex( 0.0 );
                     texcoords.push_back( tex );
                     texcoords.push_back( tex );
@@ -103,9 +112,12 @@ host_mesh_t::host_mesh_t( const tinyobj::shape_t &shape ) {
     *this = std::move( host_mesh_t( vertices, normals, texcoords, indices, texture_name, ambient_color, diffuse_color ) );
 }
 
-void host_mesh_t::update_aabb_() const {
-    if( !aabb_cached_ ) {
-        for( auto vertex : vertices ) {
+void host_mesh_t::update_aabb_() const
+{
+    if( !aabb_cached_ )
+    {
+        for( auto vertex : vertices )
+        {
             aabb_.first.x = glm::min( aabb_.first.x, vertex.x );
             aabb_.first.y = glm::min( aabb_.first.y, vertex.y );
             aabb_.first.z = glm::min( aabb_.first.z, vertex.z );
